+++ sdk/firstload/CHANGELOG.md
@@ -1,5 +1,17 @@
 # highlight.run
 
+## 9.9.1
+
+### Patch Changes
+
+-   91f82fc: record highlight session id on client-side metrics
+
+## 9.9.0
+
+### Minor Changes
+
+-   ab181de: add a network performance listener to report network connection properties
+
 ## 9.8.0
 
 ### Minor Changes